moved utility scripts from /scripts to standalone example, plus removed unnecessary deps
